Build makefile plugins if present

+++ tools/buildall-linux.sh
@@ -13,11 +13,15 @@ export CXXFLAGS="-DEL_APPIMAGE=1"
 ./waf configure --prefix=/usr $@
 ./waf resave clean build docs --progress
 
-cd "${here}/tools/jucer/Element/Builds/LinuxMakefile"
-make clean && rm -rf build && make -j4 CONFIG=Release
+if [ -f "${here}/tools/jucer/Element/Builds/LinuxMakefile" ]; then
+    cd "${here}/tools/jucer/Element/Builds/LinuxMakefile"
+    make clean && rm -rf build && make -j4 CONFIG=Release
+fi
 
-cd "${here}/tools/jucer/ElementFX/Builds/LinuxMakefile"
-make clean && rm -rf build && make -j4 CONFIG=Release
+if [ -f "${here}/tools/jucer/ElementFX/Builds/LinuxMakefile" ]; then
+    cd "${here}/tools/jucer/ElementFX/Builds/LinuxMakefile"
+    make clean && rm -rf build && make -j4 CONFIG=Release
+fi
 
 cd "${here}"
